safari-600.7-branch/LayoutTests/userscripts/user-script-video-document-expected.txt

CONSOLE MESSAGE: line 1: End Script: user-script-video-document.html
CONSOLE MESSAGE: line 1: Start Script: silence-loop.mov
CONSOLE MESSAGE: line 1: End Script: silence-loop.mov
